+++ R/BDtoPyPop.R
@@ -130,6 +130,7 @@ pypopHeaders <- function(colHead) {
 #'@param dataset A data frame containing a BIGDAWG-formatted case-control dataset
 #'@param filename A character string identifying the desired path and name for the generated files or the elements of the returned list object.
 #'@param save.file A logical that determines if a pair of files should be written (TRUE) or if a list object should be returned. The default value is TRUE.
+#'@param save.path A character string identifying the path in which to write the pair of files when save.file is TRUE. The default value is tempdir().
 #'
 #'@return When save.file = TRUE, a pair of files named "'filename'.positive.pop" and "'filename'.negative.pop" are generated in the working directory. When save.file = FALSE, a list of two-elements, named "'filename'.positive" and "'filename'.negative", is returned. 
 #'
@@ -144,7 +145,7 @@ pypopHeaders <- function(colHead) {
 #'\dontrun{
 #' HLAdata.PP <- BDtoPyPop(BIGDAWG::HLA_data,"BDHLA",FALSE)
 #'}
-BDtoPyPop <- function(dataset, filename, save.file=TRUE) {
+BDtoPyPop <- function(dataset, filename, save.file=TRUE,save.path = tempdir()) {
   
   colnames(dataset) <- pypopHeaders(colnames(dataset))
   dataset <- convertAny(dataset)
@@ -153,8 +154,8 @@ BDtoPyPop <- function(dataset, filename, save.file=TRUE) {
   negSet <- dataset[dataset[,2] %in% c(0,"0"),]
   
   if(save.file) {
-    write.table(posSet,paste(filename,"positive","pop",sep="."),append = FALSE,quote = FALSE,sep ="\t",na = "****",row.names = FALSE,col.names = TRUE)
-    write.table(negSet,paste(filename,"negative","pop",sep="."),append = FALSE,quote = FALSE,sep ="\t",na = "****",row.names = FALSE,col.names = TRUE)
+    write.table(posSet,paste(save.path,paste(filename,"positive","pop",sep="."),sep=.Platform$file.sep),append = FALSE,quote = FALSE,sep ="\t",na = "****",row.names = FALSE,col.names = TRUE)
+    write.table(negSet,paste(save.path,paste(filename,"negative","pop",sep="."),sep=.Platform$file.sep),append = FALSE,quote = FALSE,sep ="\t",na = "****",row.names = FALSE,col.names = TRUE)
   } else {
     
     retList <- list(posSet,negSet)